requirements.txt
# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Тесты
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-html>=4.0.0
httpx>=0.27.0
